+++ internal/cache/default_gcil.go
@@ -1,1 +1,3 @@
 package cache
+
+const envGolangciLintCache = "GOLANGCI_LINT_CACHE"
