+++ internal/cache/hash_gcil.go
@@ -1,1 +1,5 @@
 package cache
+
+func SetSalt(b []byte) {
+	hashSalt = b
+}
